organizing
Started to organize the mainbot code and the irc code. also made some things like fileIO and the variables for a more global variable storage system.
