+++ modules/fileIO.py
@@ -1,0 +1,13 @@
+import json
+
+#file load and save stuff
+def fileSave(fileName,config):
+    print("Saving")
+    f = open(fileName, 'w') #opens the file your saving to with write permissions
+    f.write(json.dumps(config,sort_keys=True, indent=4 ) + "\n") #writes the string to a file
+    f.close() #closes the file io
+
+def fileLoad(fileName):#loads files
+    with open(fileName, 'r') as handle:#loads the json file
+        config = json.load(handle) 
+    return config
