+++ modules/variables.py
@@ -1,0 +1,13 @@
+config = {"channelName": "", "pageToken": "", "serverName": "", "discordToken": "","discordToIRCFormating": "", "IRCToDiscordFormatting":""}
+mainMsg = []
+tempRole = 0
+discordMembers = {}
+processedCommand = []
+processedMSG = []
+discordRoles = {}#need to look futher into this for doing roles/ranks accross multiple services
+
+
+
+
+    
+
